Fix DbSet 🚀🚀🚑

+++ EventsMS/Models/Category.cs
@@ -2,7 +2,7 @@
 
 namespace EventsMS.Models;
 
-public class Category:BaseEntities.BaseEntity<int>
+public class Category:BaseEntities.BaseEntity<long>
 {
     public string Name { get; set; } = null!;
     public string? Description { get; set; }
